Migrate Store instantiation
