+++ bench/block_processing.exs
@@ -6,6 +6,7 @@ alias LambdaEthereumConsensus.Store.BlockStore
 alias LambdaEthereumConsensus.Store.Db
 alias LambdaEthereumConsensus.Store.StateStore
 alias Types.{BeaconState, SignedBeaconBlock}
+alias Types.StoreImpl
 
 {:ok, _} = Db.start_link(nil)
 Cache.initialize_cache()
@@ -19,7 +20,7 @@ IO.puts("fetching blocks...")
 {:ok, %SignedBeaconBlock{} = new_block} = BlockStore.get_block_by_slot(slot + 1)
 
 IO.puts("initializing store...")
-{:ok, store} = Helpers.get_forkchoice_store(state, block, true)
+{:ok, store} = Helpers.get_forkchoice_store(state, block, StoreImpl.InMemory)
 store = Handlers.on_tick(store, store.time + 30)
 
 attestations = new_block.message.body.attestations
