+++ lib/spec/runners/fork_choice.ex
@@ -10,6 +10,7 @@ defmodule ForkChoiceTestRunner do
   alias LambdaEthereumConsensus.ForkChoice.Helpers
   alias Types.SignedBeaconBlock
   alias Types.Store
+  alias Types.StoreImpl
 
   @disabled_on_block_cases [
     # "basic",
@@ -105,7 +106,7 @@ defmodule ForkChoiceTestRunner do
 
     signed_block = %SignedBeaconBlock{message: anchor_block, signature: <<0::768>>}
 
-    {:ok, store} = Helpers.get_forkchoice_store(anchor_state, signed_block, false)
+    {:ok, store} = Store.get_forkchoice_store(anchor_state, signed_block, StoreImpl.InMemory)
 
     assert {:ok, _store} = apply_steps(case_dir, store, steps)
   end
